+++ .npmignore
@@ -105,7 +105,7 @@ dist
 .tern-port
 
 # Level DB files
-**/db_test/*
+**/db_test
 *.zkey
 *.wasm
 *.r1cs
