+++ src/global.ts
@@ -1,42 +1,10 @@
-import bigInt from 'big-integer';
-
 import { buildPoseidon } from './crypto/poseidon_wasm.js';
 import { HashFunction } from './witnesses/fixed-merkle-tree/index.js';
 import buildEddsa from './crypto/eddsa.js';
 import { getCurveFromName } from './crypto/ffjavascript.js';
+import { Hash0, Hash1, Hasher, SnarkField, ZidenParams } from './index.js';
 
-export const SNARK_SIZE: bigInt.BigNumber = bigInt(
-  '21888242871839275222246405745257275088548364400416034343698204186575808495617'
-);
 
-export interface ZidenParams {
-  readonly hasher: Hasher;
-  readonly hash0: Hash0;
-  readonly hash1: Hash1;
-  readonly fmtHash: HashFunction;
-  readonly eddsa: EDDSA;
-  readonly F: SnarkField;
-}
-export type Hasher = (arr: Array<BigInt | ArrayLike<number>>) => ArrayLike<number>;
-export type Hash0 = (left: BigInt | ArrayLike<number>, right: BigInt | ArrayLike<number>) => ArrayLike<number>;
-export type Hash1 = (key: BigInt | ArrayLike<number>, value: BigInt | ArrayLike<number>) => ArrayLike<number>;
-export interface SnarkField {
-  toObject: (arr: ArrayLike<number>) => BigInt;
-  e: (num: BigInt | ArrayLike<number> | number | string) => ArrayLike<number>;
-  one: ArrayLike<number>;
-  zero: ArrayLike<number>;
-  eq: (value1: ArrayLike<number>, value2: ArrayLike<number>) => boolean;
-  isZero: (value: ArrayLike<number>) => boolean;
-  toString: (value: ArrayLike<number>) => string;
-}
-export interface EDDSASignature {
-  R8: Array<ArrayLike<number>>;
-  S: BigInt;
-}
-export interface EDDSA {
-  prv2pub: (privateKey: Buffer) => Array<ArrayLike<number>>;
-  signPoseidon: (privateKey: Buffer, msg: ArrayLike<number>) => EDDSASignature;
-}
 
 export async function buildHasher(): Promise<Hasher> {
   return await buildPoseidon();
