--- src/trees/claimRev/LOG
@@ -1,5 +1,0 @@
-2022/12/16-17:54:52.531414 7f1a18ecc640 Recovering log #121
-2022/12/16-17:54:52.531922 7f1a18ecc640 Level-0 table #123: started
-2022/12/16-17:54:52.626244 7f1a18ecc640 Level-0 table #123: 727 bytes OK
-2022/12/16-17:54:52.760014 7f1a18ecc640 Delete type=0 #121
-2022/12/16-17:54:52.760795 7f1a18ecc640 Delete type=3 #120
