--- src/trees/claimRev/LOG.old
@@ -1,3 +1,0 @@
-2022/12/16-17:54:06.107339 7fc795ffb640 Recovering log #119
-2022/12/16-17:54:06.195438 7fc795ffb640 Delete type=0 #119
-2022/12/16-17:54:06.227139 7fc795ffb640 Delete type=3 #118
