--- src/trees/claimRev1/CURRENT
@@ -1,1 +1,0 @@
-MANIFEST-000034
