--- src/trees/claimRev1/LOG
@@ -1,3 +1,0 @@
-2022/12/16-17:50:55.779327 7fad757fa640 Recovering log #33
-2022/12/16-17:50:55.997723 7fad757fa640 Delete type=0 #33
-2022/12/16-17:50:55.999113 7fad757fa640 Delete type=3 #32
