--- src/trees/claimRev1/LOG.old
@@ -1,3 +1,0 @@
-2022/12/16-17:49:31.008456 7f2be37fe640 Recovering log #31
-2022/12/16-17:49:31.308236 7f2be37fe640 Delete type=0 #31
-2022/12/16-17:49:31.326282 7f2be37fe640 Delete type=3 #30
