--- src/witnesses/index.ts
@@ -1,5 +1,0 @@
-export * as authentication from './authentication.js';
-export * as stateTransition from './stateTransition.js';
-export * as queryMTP from './queryMTP.js';
-export * as querySig from './querySig.js';
-export * as query from './query.js';
